+++ src/problem/Problems.kt
@@ -3,6 +3,7 @@ package problem
 import discrete.Problem
 
 const val Allocation = "allocation"
+const val Assignment = "assignment"
 const val BinCover = "bin_cover"
 const val BinPacking = "bin_packing"
 const val CarPainting = "car_painting"
@@ -43,6 +44,7 @@ const val WarehouseLocation = "warehouse_location"
 
 val Creator: Map<String, (String, Int) -> Problem?>  = mapOf(
 	Allocation to ::newAllocation,
+	Assignment to ::newAssignment,
 	BinCover to ::newBinCover,
 	BinPacking to ::newBinPacking,
 	CarPainting to ::newCarPainting,